\name{comp2col}
\alias{comp2col}
\title{Creates colors and legenda for comparisson treemaps}
\usage{
  comp2col(dat, upperboundText, showScale, neg)
}
\arguments{
  \item{dat}{}

  \item{upperboundText}{}

  \item{showScale}{}

  \item{neg}{}
}
\value{
  colorscale
}
\description{
  Creates colors and legenda for comparisson treemaps
}

