% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.functions.run.R
\name{conditional}
\alias{conditional}
\title{Perform simple stepwise conditional analysis for fine mapping.}
\usage{
conditional(
  input.folder = "./",
  pattern = ".txt",
  output.folder = "./",
  mc.cores = 1
)
}
\arguments{
\item{input.folder:}{The directory where input files are available. Default is the current working directory.}

\item{pattern:}{A string that matches part of the input file names. This string shall be present in only input files, and absent from other file names in the same folder. Default is '.txt'.}

\item{output.folder:}{The directory where output files and intermediate files are to be saved. If the specified directory does not exist, it will be automatically created. Default is the current working directory.}

\item{mc.cores:}{The number of parallel processes that will be forked for batch processing. This is useful for multi-core desktops or server clusters. Default is 1.}
}
\value{
NULL
The final fine mapping result is written to the file xx.treemap.out. The "locus" column indicates unique causal locus. The "rank" column indicates ranking of variants within each locus. Variants with rank of 1 are lead eVars at a locus. P-values and effects of lead eVars are in the "p" column and the "eff" column, respectively. Variants with rank >1 are ordered based on their linkage to the lead eVar at each locus, with r2 values shown in the "cor" column.
}
\description{
Perform simple stepwise conditional analysis for fine mapping.
}
