\name{createRec}
\alias{createRec}
\title{Creates graphical rectangle objects out of coordinates}
\usage{
  createRec(recList, filled = TRUE, label = "",
    labelbg = FALSE, labellb = 0, lwd = 1)
}
\arguments{
  \item{recList}{}

  \item{filled}{}

  \item{label}{}

  \item{labelbg}{}

  \item{labellb}{}

  \item{lwd}{export}
}
\description{
  Creates graphical rectangle objects out of coordinates
}

