\name{dens2col}
\alias{dens2col}
\title{Creates colors and legenda for density treemaps}
\usage{
  dens2col(dat, upperboundText, showScale, neg)
}
\arguments{
  \item{dat}{}

  \item{upperboundText}{}

  \item{showScale}{}

  \item{neg}{}
}
\value{
  colorscale
}
\description{
  Creates colors and legenda for density treemaps
}

