\name{fill2col}
\alias{fill2col}
\title{Creates colors and legenda for treemaps with percentages}
\usage{
  fill2col(dat, upperboundText, showScale, neg)
}
\arguments{
  \item{dat}{}

  \item{upperboundText}{}

  \item{showScale}{}

  \item{neg}{}
}
\value{
  colorscale
}
\description{
  Creates colors and legenda for treemaps with percentages
}

