\name{GNI2010}
\alias{GNI2010}
\title{GNI 2010 Data}
\description{
Gross national income (per capita) in dollars per country
in 2010.
}
\references{
\href{http://data.worldbank.org/data-catalog/health-nutrition-and-population-statistics}{Website
of The World Bank - Health Nutrition and Population
Statistics}
}

