\name{tmLocate}
\alias{tmLocate}
\title{Return info about clicked rectangle}
\usage{
  tmLocate(coor, tmSave)
}
\arguments{
  \item{coor}{mouse click in npc coordinates}

  \item{tmSave}{treemap information: value returned by
  \code{\link{treemap}}}
}
\description{
  Return information about the rectangle the user clicked
  on. Useful to identify small rectangles in large
  treemaps.
}

