\name{treecolors}
\alias{treecolors}
\title{Interactive tool to expore Tree Colors}
\usage{
treecolors(height = 700)
}
\arguments{
  \item{height}{height of the plotted treemap in pixels.
  Tip: decrease this number if the treemap doesn't fit
  conveniently.}
}
\description{
Tree Colors are color palettes for tree data structures.
They are used in \code{\link{treemap}} by default
(\code{type="index"}). With this tool, users can experiment
with the parameters (in \code{\link{treemap}} stored in
\code{palette.HCL.options}).
}
\examples{
\dontrun{
treecolors()
}
}

