% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.functions.run.R
\name{treemap}
\alias{treemap}
\title{Perform TreeMap analysis for fine mapping.}
\usage{
treemap(
  input.folder = "./",
  pattern = ".txt",
  output.folder = "./",
  steps = 1:9,
  mc.cores = 1
)
}
\arguments{
\item{input.folder:}{The directory where input files are available. Default is the current working directory. Each input file is a tab-delimited text file. The first line has column headers with no specific restrictions. The remaining lines have sample ID in the 1st column, gene expression value in the 2nd column, and genotype data in the other columns. Genotypes values are coded as 0/1/2 for homozygous reference, heterozygous and homozygous alternative, respectively. Each row has data from one individual. A sample input file "SMNT.simulated.txt" is available in the "examples/" folder.}

\item{pattern:}{A string that matches part of the input file names. This string shall be present in only input files, and absent from other file names in the same folder. Default is '.txt'.}

\item{output.folder:}{The directory where output files and intermediate files are to be saved. If the specified directory does not exist, it will be automatically created. Default is the current working directory.}

\item{steps:}{A vector of integers corresponding to the functions that will be executed. Default is 1:9. The final output is written to a file named xx.treemap.out where xx matches the name of the input file.
1: pre-processing. This step removes variants with missing genotypes in >50% samples, and remove samples with missing genotypes of >50% variants. It performs z-transformation to scale each genotype column. It groups variants into a hierarchical structure based on linkage disequilibrium. This step produces several intermediate files including xx.response.csv.gz, xx.geno.csv.gz, xx.geno.norm.csv.gz and xx.group.csv.gz that correspond to gene expression values, original genotypes, normalized genotype values, LD groups, respectively.
2: outer-layer. This step performs tree-guided group lasso. It produces an intermediate file xx.tree-lasso.csv
3: outer-layer. This step parses lasso output. It produces an intermediate file xx.top.txt.
4: middle-layer. This step selects variants within an LD block. It produces an intermediate file xx.within.RData.
5: inner-layer. This step selects variants across multiple LD blocks and individual variants. It produces an intermediate file xx.cross.RData.
6: inner-layer. This step produces the global optimal combination of multiple loci. It produces an intermediate file xx.locus.txt.
7: inner-layer. This step ranks variants within each selected locus. It produces an intermediate file xx.comb.txt.
8: estimate effect: This step esimates the effect size of each top-ranked variant at a locus via multivariate linear regression. It produces the final output file xx.treemap.out.
9: log run-time: This step writes run time of each of the previous step in a log file xx.log.txt.}

\item{mc.cores:}{The number of parallel processes that will be forked for batch processing. This is useful for multi-core desktops or server clusters. Default is 1.}
}
\value{
NULL
The final fine mapping result is written to the file xx.treemap.out. The "locus" column indicates unique causal locus. The "rank" column indicates ranking of variants within each locus. Variants with rank of 1 are lead eVars at a locus. Lead eVars are included in the global optimal solution, and their estimated effect sizes are in the "effect.est" column. Variants with rank >1 are suboptimal solutions that are linked to the lead eVar at each locus, with r2 values shown in the "cor" column.
}
\description{
Perform TreeMap analysis for fine mapping.
}
