\name{vColorDivSplit}
\alias{vColorDivSplit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	vColorDivSplit
}
\description{
	Splits string of color column names at "/".
}
\usage{
vColorDivSplit(vColor)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vColor}{
  }
}
\details{
}
\value{
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
