% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{business}
\alias{business}
\title{Fictitious  Business Statistics Data}
\description{
Fictitious (aggregated) business statistics data. The index variables (\code{NACE1} to \code{NACE4}) are derived from the Statistical Classification of Economic Activities in the European Community (NACE). The variables \code{turnover(.prev)} and \code{employees(.prev)} have values for NACE codes in the business economy domain only.
}
\references{
\href{https://en.wikipedia.org/wiki/Statistical_Classification_of_Economic_Activities_in_the_European_Community}{Statistical Classification of Economic Activities in the European Community (NACE)}
\href{https://ec.europa.eu/eurostat/web/structural-business-statistics}{Eurostat - Structural business statistics (SBS)}
}
