% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{GNI2014}
\alias{GNI2014}
\title{GNI 2014 Data}
\description{
Gross national income (per capita) in dollars and population totals per country in 2014.
}
\details{
The GNI numbers from the World Bank are based on the Atlas. The population data are taken from Natural Earth Data.
}
\references{
\href{https://datacatalog.worldbank.org/dataset/gni-capita-ranking-atlas-method-and-ppp-based}{The World Bank - GNI per capita ranking}
\href{https://www.naturalearthdata.com/}{Natural Earth Data}
}
