% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeList}
\alias{makeList}
\title{Function to make the list}
\usage{
makeList(x)
}
\arguments{
\item{x}{table containing}
}
\value{
list of tree nodes
}
\description{
Given a table in input a list is created to generate the tree nodes
}
\examples{
# makelist(x)
}
\author{
Claudia Vitolo
}

