\name{tmLocate}
\alias{tmLocate}
\title{Return info about clicked rectangle}
\usage{
  tmLocate(npcClick, tmSave)
}
\arguments{
  \item{npcClick}{mouse click in npc coordinates, which can
  be generated by \code{\link{tmClick}}}

  \item{tmSave}{treemap information: value returned by
  \code{\link{treemap}}}
}
\description{
  Return information about the rectangle the user clicked
  on. Useful to identify small rectangles in large
  treemaps.
}
\examples{
data(business)
\dontrun{
tmSave <- treemap(business,
index=c("NACE1", "NACE2", "NACE3"),
vSize="turnover",
type="index")
# capture mouseclick
npcClick <- tmClick()

# locate clicked object
print(tmLocate(npcClick, tmSave))
}
}

