% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmPlot.R
\name{tmPlot}
\alias{tmPlot}
\title{Create a treemap (deprecated)}
\usage{
tmPlot(...)
}
\arguments{
\item{...}{passed on to \code{\link{treemap}}}
}
\description{
This function is migrated to \code{\link{treemap}}.
}
