% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecolors.R
\name{treecolors}
\alias{treecolors}
\title{Interactive tool to experiment with Tree Colors}
\usage{
treecolors(height = 700)
}
\arguments{
\item{height}{height of the plotted treemap in pixels. Tip: decrease this number if the treemap doesn't fit conveniently.}
}
\description{
Tree Colors are color palettes for tree data structures. They are used in \code{\link{treemap}} by default (\code{type="index"}). With this tool, users can experiment with the parameters (in \code{\link{treemap}} stored in \code{palette.HCL.options}). Tree Colors can directly be obtained by \code{\link{treepalette}} with \code{method="HCL"}.
}
\examples{
\dontrun{
treecolors()
}
}
