\docType{package}
\name{treemap-package}
\alias{treemap-package}
\title{Treemap package}
\description{
\tabular{ll}{ Package: \tab treemap\cr Type: \tab
Package\cr Version: \tab 2.1-1\cr Date: \tab 2013-12-05\cr
License: \tab GPL-3\cr LazyLoad: \tab yes\cr }
}
\details{
A treemap is a space-filling visualization of hierarchical
structures. This package offers great flexibility to draw
treemaps.

The main function is \code{\link{treemap}}. See also
\code{\link{itreemap}} for a graphical user interface.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{treemap}

