% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{treemap-package}
\alias{treemap-package}
\title{Treemap package}
\description{
\tabular{ll}{
Package: \tab treemap\cr
Type: \tab Package\cr
Version: \tab 2.4\cr
Date: \tab 2015-10-14\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\details{
A treemap is a space-filling visualization of hierarchical structures. This package offers great flexibility to draw treemaps.

The main function is \code{\link{treemap}}. See also \code{\link{itreemap}} for a graphical user interface to create treemaps. By default Tree Colors are used, which are colors from the HCL color space. Use \code{\link{treecolors}} to experiment with the parameter settings.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{treemap}

