\docType{package}
\name{treemap-package}
\alias{treemap-package}
\alias{treemap-treemap}
\title{Treemap visualization}
\description{
  \tabular{ll}{ Package: \tab treemap\cr Type: \tab
  Package\cr Version: \tab 1.1-1\cr Date: \tab
  2012-07-10\cr License: \tab GPL-3\cr LazyLoad: \tab
  yes\cr }
}
\details{
  A treemap is a space-filling visualization of
  hierarchical structures. This package offers great
  flexibility to draw treemaps.

  See \code{\link{treemap}}
}
\author{
  Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{treemap}

