% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{treemap-package}
\alias{treemap-package}
\title{Treemap package}
\description{
A treemap is a space-filling visualization of hierarchical structures. This package offers great flexibility to draw treemaps.
}
\details{
The main function is \code{\link{treemap}}. See also \code{\link{itreemap}} for a graphical user interface to create treemaps. By default Tree Colors are used, which are colors from the HCL color space. Use \code{\link{treecolors}} to experiment with the parameter settings.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{treemap}
